[FLINK-4705] Instrument FixedLengthRecordSorter
Updates comparators with support for key normalization.

+++ flink-core/src/test/java/org/apache/flink/api/java/typeutils/runtime/TupleComparatorILDXC2Test.java
@@ -51,7 +51,7 @@ public class TupleComparatorILDXC2Test extends TupleComparatorTestBase<Tuple3<In
 					new DoubleComparator(ascending),
 					new LongComparator(ascending)
 				},
-		new TypeSerializer[]{ IntSerializer.INSTANCE, DoubleSerializer.INSTANCE, LongSerializer.INSTANCE });
+		new TypeSerializer[]{ IntSerializer.INSTANCE, LongSerializer.INSTANCE, DoubleSerializer.INSTANCE });
 	}
 
 	@SuppressWarnings("unchecked")
